+++ main.ts
@@ -1,0 +1,6 @@
+import './src/init/import.js';
+import {Application} from './src/application/app.js';
+
+const app = new Application();
+
+//notification!
