+++ src/init/dotenv.ts
@@ -1,0 +1,3 @@
+import _env from 'dotenv';
+
+_env.config();
