+++ src/init/import.ts
@@ -1,0 +1,3 @@
+import './dotenv.js';
+import './node_extension.js';
+
