--- src/interfaces/commands.ts
@@ -1,7 +1,0 @@
-import { SlashCommandBuilder } from "discord.js";
-
-export default interface CommandsInterface {
-    event: string,
-	data: SlashCommandBuilder;
-	execute: () => any;
-}
