--- src/interfaces/events.ts
@@ -1,7 +1,0 @@
-import { ClientEvents } from "discord.js";
-
-export default interface EventsInterface {
-    event: ClientEvents,
-	once: boolean;
-	execute: () => any;
-}
