+++ src/resources/all.ts
@@ -1,0 +1,3 @@
+export * from './intefaces.js';
+export * from './enums.js';
+export default {};
