+++ src/resources/enums.ts
